+++ NEWS.md
@@ -1135,6 +1135,9 @@ Changes to the Mercury compiler
 * The `--no-ansi-c` option has been deprecated and will be removed in a future
   release.  The option has not had any effect for a long time.
 
+* We have deleted the `--java-object-file-extension` option, which has
+  only ever had one useful value, which is the default value, ".class".
+
 Changes to the Mercury debugger
 -------------------------------